EbisBank Future is Now
We are offering something that no one will do, sovereignty data privacy where individuals and government will be the sole manager of the information, and experience safer, faster, and most comfortable facilities place to join the next generation of digital community where humanity is the real central interest. first patent digital genomic onboarding identity, second patent, it will be linked to the first with a unique real decentralized finance with will give access to whom currently cannot and increase the experience of those who already have an identity.